refactor: Se cambian tipos de datos de DTO y entidades para asociar a la BD

+++ tienda/src/main/java/com/mintic/tienda/entity/Cliente.java
@@ -15,20 +15,23 @@ public class Cliente {
     @Id
     @GeneratedValue( strategy=GenerationType.AUTO )
     @Column(name = "id_cliente")
-    private int idCliente;
+    private Long idCliente;
 
     @Column(name = "nombre")
     private String nombre;
 
-    public int getIdCliente() {
+    public Long getIdCliente() {
         return idCliente;
     }
-    public void setIdCliente(int idCliente) {
+
+    public void setIdCliente(Long idCliente) {
         this.idCliente = idCliente;
     }
+
     public String getNombre() {
         return nombre;
     }
+    
     public void setNombre(String nombre) {
         this.nombre = nombre;
     }
