+++ tienda/src/main/java/com/mintic/tienda/repository/ITipoPrendaRepo.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.mintic.tienda.entity.TipoPrenda;
 
 @Repository
-public interface ITipoPrendaRepo extends JpaRepository<TipoPrenda, Integer>{
+public interface ITipoPrendaRepo extends JpaRepository<TipoPrenda, Long>{
     
 }
