+++ tienda/src/main/java/com/mintic/tienda/service/DTO/TallaDTO.java
@@ -3,18 +3,21 @@ package com.mintic.tienda.service.DTO;
 
 public class TallaDTO {
     
-    private int idTalla;
+    private Long idTalla;
     private String talla;
 
-    public int getIdTalla() {
+    public Long getIdTalla() {
         return idTalla;
     }
-    public void setIdTalla(int idTalla) {
+
+    public void setIdTalla(Long idTalla) {
         this.idTalla = idTalla;
     }
+
     public String getTalla() {
         return talla;
     }
+    
     public void setTalla(String talla) {
         this.talla = talla;
     }
